[breadboard-extension] Move transpilation out
